[FVM] Removing meter.Meter interface
